Updated plant function. Changed ints to doubles

+++ plot.py
@@ -7,8 +7,8 @@ y = []
 with open('out.csv', 'r') as csvfile:
     data = csv.reader(csvfile, delimiter=',')
     for row in data:
-        x.append(int(row[0]))
-        y.append(int(row[1]))
+        x.append(float(row[0]))
+        y.append(float(row[1]))
 
 plt.plot(x, y)
 plt.show();
